159 - Create shared styles for cards
